Move writeinpublic views and urls into app
